[#62] feat: 로그아웃 구현 완료
