+++ src/main/java/com/lovecloud/global/config/RedisConfig.java
@@ -13,13 +13,13 @@ import org.springframework.data.redis.serializer.StringRedisSerializer;
 @Configuration
 public class RedisConfig {
 
-    @Value("${spring.data.redis.host}")
+    @Value("${spring.redis.data.host}")
     private String host;
 
-    @Value("${spring.data.redis.port}")
+    @Value("${spring.redis.data.port}")
     private String port;
 
-    @Value("${spring.data.redis.password:}")
+    @Value("${spring.redis.data.password}")
     private String password;
 
     @Bean
